Ported all the common concerns into .NET Standard project
