+++ PocketWallet.Bkash.Common/PocketWallet.Bkash.Common.csproj
@@ -9,4 +9,8 @@
     <InternalsVisibleTo Include="PocketWallet.Bkash" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="System.Text.Json" Version="9.0.10" />
+  </ItemGroup>
+
 </Project>
