+++ PocketWallet.Bkash.Common/Providers/TokenProvider.cs
@@ -1,0 +1,46 @@
+﻿using PocketWallet.Bkash.Common.Abstractions;
+using System;
+
+namespace PocketWallet.Bkash.Common.Providers
+{
+    /// <summary>
+    /// Represents token storage mechanism tool.
+    /// </summary>
+    internal class TokenProvider : ITokenProvider
+    {
+        private readonly IDateTimeProvider _timeProvider;
+
+        private DateTime _expiry;
+        private string _accessToken = string.Empty;
+        private string _refreshToken = string.Empty;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="TokenProvider"/> class.
+        /// </summary>
+        /// <param name="timeProvider">Date and time provider instance.</param>
+        public TokenProvider(IDateTimeProvider timeProvider)
+        {
+            _timeProvider = timeProvider;
+        }
+
+        /// <inheritdoc/>
+        public string AccessToken => _accessToken;
+
+        /// <inheritdoc/>
+        public string RefreshToken => _refreshToken;
+
+        /// <inheritdoc/>
+        public bool IsTokenAvailable => !string.IsNullOrWhiteSpace(_accessToken);
+
+        /// <inheritdoc/>
+        public bool IsTokenAlive => IsTokenAvailable && _expiry > _timeProvider.UtcNow;
+
+        /// <inheritdoc/>
+        public void Set(string accessToken, string refreshToken, DateTime expiry)
+        {
+            _accessToken = accessToken;
+            _refreshToken = refreshToken;
+            _expiry = expiry;
+        }
+    }
+}
