+++ PocketWallet.Bkash/GlobalUsings.cs
@@ -1,4 +1,3 @@
-﻿global using System.Text.Json.Serialization;
-global using PocketWallet.Bkash.Abstraction;
-global using PocketWallet.Bkash.Models;
+﻿global using PocketWallet.Bkash.Abstraction;
+global using PocketWallet.Bkash.Common.Models;
 global using CONSTANTS = PocketWallet.Bkash.Common.Constants.Constants;
