--- PocketWallet.Bkash/Models/BkashTokenRequest.cs
@@ -1,19 +1,0 @@
-﻿namespace PocketWallet.Bkash.Models;
-
-/// <summary>
-/// Represents Bkash token request object.
-/// </summary>
-internal class BkashTokenRequest
-{
-    /// <summary>
-    /// Application Key value shared by Bkash during on-boarding.
-    /// </summary>
-    [JsonPropertyName("app_key")]
-    internal string AppKey { get; init; } = string.Empty;
-
-    /// <summary>
-    /// Application Secret value shared by Bkash during on-boarding.
-    /// </summary>
-    [JsonPropertyName("app_secret")]
-    internal string AppSecret { get; init; } = string.Empty;
-}
